Add files via upload
Equipment Management System, based on a warehouse, on this website written in HTML, CSS, JavaScript you can manage users, locations and equipment. There are two types of Admin and User profiles, the user can select equipment and locations that will be used and where they will use it, going through a mandatory list to check if the equipment is in working order, and can also report damage to equipment leaving it. or unavailable for use. Admin: will have complete access to users and can create, modify and have full control over users, with times, dates and history of when changes or actions were made to users, equipment and locations.

You can see examples of the site in the photos preview.png

I spent 6 months learning how to build this idea from scratch without any knowledge, today I make it available on my github as a way to show that I am capable of creating solutions for companies.